添加拓扑配置; 完善NDP功能。 NDP功能待完整测试。

+++ protocol/buffer.go
@@ -4,6 +4,6 @@ type Buffer struct {
 	Octet []byte
 }
 
-func (buf *Buffer) Length() int {
-	return len(buf.Octet)
+func (buf *Buffer) Length() uint16 {
+	return uint16(len(buf.Octet))
 }
